// src/app/modules/master/models/articles.ts
export class ArticleData {
        category: string;
        title: string;
        content: string;
        article_image: string;
        title_order: number;
        unique_tag: string;
        foot_text_flg: string;
        foot_text_content: string;
        foot_text_redirect: string;
        foot_image_flg: string;
        foot_image_content: string;
        foot_image_redirect: string;
        foot_button_redirect: string;
        foot_button_flg: string;
        foot_button_content: string;
}